Add mouse wheel menu option
Currently the mouse wheel can be used to switch tabs in a QTabBar what can be very annoying for touchpad users with multitouch scroll/wheel options. There are considerations to remove this currently for mac users https://codereview.qt-project.org/#/c/109523/ . As a quick fix a checkbox is added to the context menu to disable the mouse wheel.